move ir_replace from cinn/optim to cinn/ir/utils
