+++ paddle/cinn/ir/utils/CMakeLists.txt
@@ -9,4 +9,5 @@ gather_srcs(
   ir_verify.cc
   ir_compare.cc
   ir_nodes_collector.cc
-  ir_copy.cc)
+  ir_copy.cc
+  ir_replace.cc)
