+++ paddle/cinn/optim/unroll_loops.cc
@@ -21,7 +21,7 @@
 #include "paddle/cinn/ir/utils/ir_copy.h"
 #include "paddle/cinn/ir/utils/ir_mutator.h"
 #include "paddle/cinn/ir/utils/ir_printer.h"
-#include "paddle/cinn/optim/ir_replace.h"
+#include "paddle/cinn/ir/utils/ir_replace.h"
 
 namespace cinn {
 namespace optim {
@@ -95,7 +95,7 @@ struct UnrollMutator : public ir::IRMutator<Expr*> {
     for (int i = min->value; i < extent->value; i++) {
       Expr start = op->min + i;
       body.push_back(optim::IRCopy(op->body));
-      optim::IrReplace(&body.back(), op->loop_var, start);
+      cinn::ir::ir_utils::IrReplace(&body.back(), op->loop_var, start);
     }
 
     *expr = ir::Block::Make(body);
